Changes made to resolve bug in case of defective return
Hello Sir,
Yesterday i found a bug in case of defective return. There are 3 extra
field for defective return but due to wrong condition written in java
file we are not able to fetch them. I have made changes to resolve this
bug. Now we are able to fetch all details in case of defective return.